refactor(56): 엔티티 자료형 변경 (#58)
* refactor: 휴가 날짜는 LocalDateTime, 휴가 일수는 Double로 변경

* test : 타입변경에 따른 테스트 수정

* test : 타입변경에 따른 테스트 수정

---------

+++ backend/team6/src/main/java/programmers/team6/domain/vacation/entity/VacationInfo.java
@@ -23,10 +23,10 @@ public class VacationInfo extends BaseEntity {
 	private int vacationId;
 
 	@Getter
-	private int totalCount;
+	private double totalCount;
 
 	@Getter
-	private int useCount;
+	private double useCount;
 
 	@Getter
 	private String vacationType;
@@ -38,7 +38,7 @@ public class VacationInfo extends BaseEntity {
 	@Version
 	private int version;
 
-	public VacationInfo(int totalCount, int useCount, String vacationType, Long memberId) {
+	public VacationInfo(double totalCount, double useCount, String vacationType, Long memberId) {
 		this.totalCount = totalCount;
 		this.useCount = useCount;
 		this.vacationType = vacationType;
@@ -47,7 +47,7 @@ public class VacationInfo extends BaseEntity {
 	}
 
 	@CheckReturnValue
-	public VacationInfoUpdateResult updateTotalCount(Integer version, Integer totalCount) {
+	public VacationInfoUpdateResult updateTotalCount(Integer version, double totalCount) {
 		if (!isSameVersion(version)) {
 			return VacationInfoUpdateResult.MISS_VERSION;
 		}
@@ -58,7 +58,7 @@ public class VacationInfo extends BaseEntity {
 		return VacationInfoUpdateResult.SUCCESS;
 	}
 
-	private boolean isOverUseCount(int totalCount) {
+	private boolean isOverUseCount(double totalCount) {
 		return this.useCount > totalCount;
 	}
